.gitlab-ci/bin/requirements.txt
colorama==0.4.5
python-gitlab==3.5.0
